refactor: extract rest routes and business logic
